Add indefensible hack for making base_url end with a slash
The reason is that some Trac-based bug trackers have, in the data
provided by oh-mainline, URLs like: http://example.com/trac and
we expect it to be http://example.com/trac/.

In the long run, what we should do is visit the Trac instance to
detect a redirect, and update the base_url that we use when we crawl,
and emit some kind of "data quality notes" that tell the user to
improve the data they give us.

Hopefully this hack will be fine and we'll never have to write that,
though.